fix: Level을 Double로 바꾸면서 발견된 코드 오류 수정

+++ suldaksuldak-core/src/main/java/com/sulsul/suldaksuldak/repo/auth/UserRepositoryCustom.java
@@ -27,7 +27,7 @@ public interface UserRepositoryCustom {
             Integer startYear,
             Integer endYear,
             Registration registration,
-            List<Integer> levelList,
+            List<Double> levelList,
             List<Integer> warningCntList,
             Boolean isActive
     );
